# Conflicts:
#	project/ovp/frontend/ovp/components/govern/glossary/glossary-info.vue
#	project/ovp/frontend/ovp/components/govern/glossary/glossary-table.vue

+++ project/ovp/frontend/ovp/components/govern/glossary/glossary-info.vue
@@ -22,6 +22,7 @@
           id="title-modify"
           class="text-input w-4/5"
           v-model="editData.name"
+          maxlength="20"
         />
         <div class="h-group gap-1">
           <button
@@ -249,9 +250,17 @@ function syncEditDataWithGlossary(): void {
   editData.description = store.glossary.description;
 }
 
-async function removeGlossary(): Promise<void> {
-  await deleteGlossary(glossary.id);
-  await getGlossaries();
+async function removeGlossary() {
+  if (confirm("데이터모델을 삭제 하시겠습니까?")) {
+    await deleteGlossary(glossary.id);
+    await getGlossaries()
+      .then(() => {
+        alert("삭제되었습니다.");
+      })
+      .catch((error) => {
+        console.error("삭제 중 오류 발생: ", error);
+      });
+  }
 }
 
 async function updateGlossary(
